Use TS and seperate test files
