+++ jest.config.ts
@@ -1,9 +1,6 @@
-/*
- * For a detailed explanation regarding each configuration property and type check, visit:
- * https://jestjs.io/docs/configuration
- */
-
 export default {
   clearMocks: true,
   coverageProvider: "v8",
+  preset: "ts-jest",
+  testEnvironment: "node",
 };
