+++ package.json
@@ -24,16 +24,16 @@
     "eslint": "^8.48.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": ">= 2.16.0",
+    "eslint-plugin-jest": "^27.2.0",
     "eslint-plugin-jsdoc": ">= 4.1.0",
     "eslint-plugin-react": ">= 7.12.4",
     "eslint-plugin-sort-destructure-keys": "^1.5.0",
     "eslint-plugin-sort-keys-fix": "^1.1.2",
-    "eslint-plugin-storybook": "^0.6.12",
-    "eslint-plugin-jest": "^27.2.0"
+    "eslint-plugin-storybook": "^0.6.12"
   },
   "devDependencies": {
     "@emotion/eslint-plugin": "^11.11.0",
-    "@types/jest": "^29.5.4",
+    "@types/jest": "^29.5.11",
     "@types/node": "^20.5.9",
     "@types/react": "^18.2.21",
     "@types/react-dom": "^18.2.7",
@@ -49,10 +49,11 @@
     "eslint-plugin-sort-destructure-keys": "^1.5.0",
     "eslint-plugin-sort-keys-fix": "^1.1.2",
     "eslint-plugin-storybook": "^0.6.13",
-    "jest": "^29.6.4",
+    "jest": "^29.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.2"
   }
 }
