+++ test/react/react.test.ts
@@ -1,0 +1,17 @@
+import { getResultJson } from "../helpers";
+
+describe("React config", () => {
+  test("That the React 'valid' file passes", async () => {
+    const [result] = await getResultJson(["test/react/valid.tsx"]);
+
+    expect(result.messages).toMatchSnapshot();
+    expect(result.errorCount).toBe(0);
+  });
+
+  test("That the react 'invalid' file throws lint errors", async () => {
+    const [result] = await getResultJson(["test/react/invalid.tsx"]);
+
+    expect(result.messages).toMatchSnapshot();
+    expect(result.errorCount).toBe(20);
+  });
+});
